requirements.txt
Pillow
opencv-python
pynput